Refactor -- moved all modal actions to individual file, delete unnecessary action state changes, and add web url validation action
