+++ client/redux/actionCreators/logout-actions.js
@@ -4,14 +4,10 @@ export const LOGOUT_FAILURE = 'LOGOUT_FAILURE';
 
 const requestLogout = () => ({
   type: LOGOUT_REQUEST,
-  isFetching: true,
-  isAuthenticated: true,
 });
 
 const receiveLogout = () => ({
   type: LOGOUT_SUCCESS,
-  isFetching: false,
-  isAuthenticated: false,
 });
 
 export const logoutUser = () => dispatch => {
