+++ client/redux/actionCreators/scenario-actions.js
@@ -1,48 +1,26 @@
+import { receiveLogin } from './login-actions';
+import { showScenarioModal } from './modal-actions';
+
 export const GET_SCENARIOS = 'GET_SCENARIOS';
 export const VALID_SCRIPT = 'VALID_SCRIPT';
 export const RESET_ATTEMPT_CHECK = 'RESET_ATTEMPT_CHECK';
-export const TOGGLE_SCENARIO_MODAL = 'TOGGLE_SCENARIO_MODAL';
 export const CURRENT_SCENARIO_ID = 'CURRENT_SCENARIO_ID';
 export const CURRENT_SPAWNS_COUNT = 'CURRENT_USER_COUNT';
 const parser = require('../../middleware/parser.js');
-
-/* ******** Scenario Modal Actions  ******** */
-export const showScenarioModal = () => ({
-  type: TOGGLE_SCENARIO_MODAL,
-  isScenarioModalOpen: true,
-});
-
-export const hideScenarioModal = () => ({
-  type: TOGGLE_SCENARIO_MODAL,
-  isScenarioModalOpen: false,
-});
-
-export const openScenarioModal = () => dispatch => {
-  dispatch(showScenarioModal());
-};
-
-export const closeScenarioModal = () => dispatch => {
-  dispatch(hideScenarioModal());
-};
-
 
 /* ******* Script Validation Actions ******* */
 export const validScript = () => ({
   type: VALID_SCRIPT,
   isValidScript: true,
-  attemptedCheck: true,
 });
 
 export const invalidScript = () => ({
   type: VALID_SCRIPT,
   isValidScript: false,
-  attemptedCheck: true,
 });
 
 export const resetCheck = () => ({
   type: RESET_ATTEMPT_CHECK,
-  isValidScript: false,
-  attemptedCheck: false,
 });
 
 export const resetAttempt = () => dispatch => dispatch(resetCheck());
@@ -71,7 +49,7 @@ export const checkValidScript = script => {
 /* ******** Scenario Get and Create Actions  ******** */
 export const allScenarios = res => ({
   type: GET_SCENARIOS,
-  scenario: JSON.parse(res),
+  scenario: JSON.parse(res.scenarios),
 });
 
 export const storeRecentScenarioID = (scenarioID) => ({
@@ -97,8 +75,10 @@ export const getScenarios = () => {
     fetch('/api/scenarios', config)
       .then(response => response.json()
         .then(res => {
-          // console.log(JSON.parse(res));
+          // Store all scenarios associated with user to state
           dispatch(allScenarios(res));
+          // Store users site token, so page refreshes will still send it back
+          dispatch(receiveLogin(res));
         })
       )
       .catch(err => console.log('Error: ', err));
@@ -122,6 +102,28 @@ export const createScenario = data => {
           dispatch(storeRecentUserCount(res.spawnsCount));
           dispatch(resetCheck());
           dispatch(showScenarioModal());
+        })
+      )
+      .catch(err => console.log('Error: ', err));
+};
+
+
+/* ******** Website Url Validation Actions  ******** */
+export const checkForValidUrl = (url, scenarioID) => {
+  const config = {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+      Authorization: `Bearer ${localStorage.getItem('id_token')}` },
+      body: `url=${url}&scenarioID=${scenarioID}`,
+  };
+
+  return dispatch =>
+    fetch('/api/validate-website', config)
+      .then(response => response.json()
+        .then(res => {
+          console.log('Response from the post request', res);
+          dispatch(getScenarios());
         })
       )
       .catch(err => console.log('Error: ', err));
